requirements.txt
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.3.7
aiohttp==3.10.4
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
bitarray==3.0.0
blinker==1.8.2
boto3==1.35.37
botocore==1.35.37
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
crc16==0.1.1
ed25519==1.5
exceptiongroup==1.2.2
fastapi==0.112.1
fastapi-storages==0.3.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
magic-filter==1.0.12
MarkupSafe==2.1.5
mnemonic==0.21
multidict==6.0.5
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
redis==5.1.1
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
sqladmin==0.18.0
SQLAlchemy==2.0.31
starlette==0.38.2
ton==0.26
tonsdk==1.0.15
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.21.0
Werkzeug==3.0.4
WTForms==3.1.2
yarl==1.9.4
